proof of logic
I separated the problems I was encountering to work on them one at a time and then merge them later. So in this case I added a sorter file to sort the subjects in order on most credits to least, and sort the results of each subject from most to least. Then I added a basic calculator that would let u delete and merge different subjects if u had more that 5 subjects. It also automatically removes the lowest rank score standards if a subject has more than 24 credits