Better results
I added a dict with all UE approved subjects so that it only returns subjects which would count for rank score. I made the dict from An XSL file on nzqa. The advantage of doing like I did is that the standards also get split into the correct subjects correctly